+++ gtm/consentio-tag-cookies/src/sandbox.js
@@ -1,0 +1,5 @@
+const log = require('logToConsole');
+log('Consentio Tag: Cookies =', data);
+
+// Variables must return a value.
+return data.cookies;